refactor: renamed main.tsx to index.tsx
