+++ src/App.tsx
@@ -1,7 +1,7 @@
-import { Footer } from './components/Footer';
-import { NavigationBar } from './components/NavigationBar';
+import { Footer } from './components/Footer/main';
+import { NavigationBar } from './components/NavigationBar/main';
 import './App.css';
-import { BloomMain } from './components/BloomMain';
+import { BloomMain } from './components/BloomMain/main';
 function App() {
   return (
     <div>
